installer-gen: new lower module owning JDK detection (moved from website-gen)
Per the agreed installer-epic layout: :installer-gen is the lower build-tooling
module; :website-gen now depends on it.

- Move JDK detection + shared HTTP/cache infra out of :website-gen into
  :installer-gen (package com.jonnyzzz.mcpSteroid.installer): Cache, PgpVerifier,
  JdkModelMain, and the model. generateJdkModel moves here too (cache dir
  gradleUserHome/caches/mcp-steroid/installer-gen-jdk, still outside build/).
- Split the model by vendor: JdkModel.kt (shared types + findJavaHome +
  resolveAllJdks orchestrator), CorrettoJdk.kt (resolveCorrettoJdks), AzulJdk.kt
  (resolveAzulJdk).
- Trim to the 5 platforms the installer ships (Corretto linux-x64/arm64,
  macos-arm64, windows-x64 + Azul windows-arm64). Drop alpine (musl: the IDEs
  need glibc; install.sh will detect-and-fail) and macos-x64. Remove the
  ALPINE_LINUX JdkOs.
- :website-gen depends on :installer-gen and reuses its KtorHttpFetcher
  (WebsiteArtifacts drops its duplicate Ktor client); its only project() dep,
  and :installer-gen has no IntelliJ deps so generateWebsite stays fast.
- CLAUDE.md guides: new installer-gen/CLAUDE.md; website-gen/CLAUDE.md trimmed to
  website artifacts; root sub-folder table updated. CI guard now asserts BOTH
  :installer-gen:test and :website-gen:test stay in ciBuildPluginTests.

Both modules compile + their hermetic suites pass. The install.sh/install.ps1
generator + Docker integration tests land next.
