+++ build.gradle.kts
@@ -279,14 +279,18 @@ val ciBuildPluginTests by tasks.registering {
             "or drop them from pluginCoreSubprojects."
     }
 
-    // :website-gen is build-tooling (website artifacts + the JDK data model; no IntelliJ deps), so it is
-    // not "plugin core". But its on-disk Cache exercises atomic file moves / path handling that genuinely
-    // benefit from the per-OS matrix, so it rides this aggregator (swept in by auto-discovery — it is in
-    // neither exclusion set) rather than getting a dedicated config. Assert that explicitly so a future
-    // refactor of the exclusion sets cannot silently drop :website-gen:test from CI.
-    require(":website-gen:test" in testTaskPaths) {
-        "ciBuildPluginTests no longer includes :website-gen:test. If :website-gen was intentionally " +
-            "excluded, give it dedicated CI coverage; otherwise keep it out of nonPluginTestSubprojects."
+    // :installer-gen + :website-gen are build-tooling (JDK detection, install-script generation, website
+    // artifacts; no IntelliJ deps), so they are not "plugin core". But :installer-gen's on-disk Cache
+    // exercises atomic file moves / path handling that genuinely benefit from the per-OS matrix, so they
+    // ride this aggregator (swept in by auto-discovery — neither is in an exclusion set) rather than
+    // getting a dedicated config. Assert it explicitly so a future refactor of the exclusion sets cannot
+    // silently drop their :test from CI. (The Docker installerIntegrationTest is a separate source set,
+    // NOT swept in here.)
+    listOf(":installer-gen:test", ":website-gen:test").forEach { required ->
+        require(required in testTaskPaths) {
+            "ciBuildPluginTests no longer includes $required. If that module was intentionally excluded, " +
+                "give it dedicated CI coverage; otherwise keep it out of nonPluginTestSubprojects."
+        }
     }
 
     dependsOn(testTaskPaths)
