+++ installer-gen/src/main/kotlin/com/jonnyzzz/mcpSteroid/installer/Cache.kt
@@ -1,5 +1,5 @@
 /* Copyright 2025-2026 Eugene Petrenko (mcp@jonnyzzz.com); Copyright 2025-2026 JetBrains. Use of this source code is governed by the Apache 2.0 license. */
-package com.jonnyzzz.mcpSteroid.websitegen
+package com.jonnyzzz.mcpSteroid.installer
 
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.cio.CIO
