+++ installer-gen/src/test/kotlin/com/jonnyzzz/mcpSteroid/installer/CacheTest.kt
@@ -1,5 +1,5 @@
 /* Copyright 2025-2026 Eugene Petrenko (mcp@jonnyzzz.com); Copyright 2025-2026 JetBrains. Use of this source code is governed by the Apache 2.0 license. */
-package com.jonnyzzz.mcpSteroid.websitegen
+package com.jonnyzzz.mcpSteroid.installer
 
 import kotlinx.serialization.Serializable
 import org.junit.jupiter.api.Test
